feat: factory-session-load-failure-debuggability-prd-002 - Return structured API diagnostics for broken factory config loads
