--- pkg/api/server_provider_sessions_test.go
@@ -1,430 +1,0 @@
-package api
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/petri"
-	"github.com/portpowered/infinite-you/pkg/testutil"
-	"go.uber.org/zap"
-	"go.uber.org/zap/zaptest/observer"
-)
-
-func TestParseCodexSessionSummary_ExtractsDiagnosticDetails(t *testing.T) {
-	session := strings.Join([]string{
-		`{"timestamp":"2026-05-18T10:00:00Z","type":"turn_context"}`,
-		`{"timestamp":"2026-05-18T10:00:01Z","type":"response_item","payload":{"type":"reasoning","summary":["checked input"],"encrypted_content":"sealed"}}`,
-		`{"timestamp":"2026-05-18T10:00:02Z","type":"response_item","payload":{"type":"function_call","call_id":"call-1","name":"exec_command","arguments":{"cmd":"go test ./pkg/api"}}}`,
-		`{"timestamp":"2026-05-18T10:00:03Z","type":"response_item","payload":{"type":"function_call_output","call_id":"call-1","output":"ok"}}`,
-		`{"timestamp":"2026-05-18T10:00:04Z","type":"event_msg","payload":{"type":"token_count","info":{"total_token_usage":{"input_tokens":100,"cached_input_tokens":40,"output_tokens":25,"reasoning_output_tokens":5,"total_tokens":130}}}}`,
-		`{"timestamp":"2026-05-18T10:00:05Z","type":"turn_context"}`,
-		`{"timestamp":"2026-05-18T10:00:06Z","type":"response_item","payload":{"type":"custom_tool_call","call_id":"call-2","name":"apply_patch","input":"patch text","status":"in_progress"}}`,
-		`{"timestamp":"2026-05-18T10:00:07Z","type":"event_msg","payload":{"type":"new_future_event"}}`,
-		`{"timestamp":"2026-05-18T10:00:08Z","type":"unexpected_top_level"}`,
-		`{bad json`,
-	}, "\n")
-
-	summary, err := parseCodexSessionSummary(strings.NewReader(session))
-	if err != nil {
-		t.Fatalf("parse codex session summary: %v", err)
-	}
-	assertCodexSessionSummaryCoreCounts(t, summary)
-	assertCodexSessionSummaryFunctionCalls(t, summary)
-	assertCodexSessionSummaryReasoning(t, summary)
-	parsed, err := parseCodexSessionDetails(strings.NewReader(session))
-	if err != nil {
-		t.Fatalf("parse codex session details: %v", err)
-	}
-	assertParsedCodexSessionSummaryTranscript(t, parsed)
-	assertCodexSessionSummaryTokenUsage(t, summary)
-	assertCodexSessionSummaryUnknowns(t, summary)
-}
-
-func TestParseCodexSessionDetails_EmitsMixedTranscriptChronologically(t *testing.T) {
-	session := strings.Join([]string{
-		`{"timestamp":"2026-05-18T10:00:00Z","type":"turn_context"}`,
-		`{"timestamp":"2026-05-18T10:00:01Z","type":"response_item","payload":{"type":"message","role":"user","content":[{"type":"input_text","text":"Inspect the failing run."}]}}`,
-		`{"timestamp":"2026-05-18T10:00:02Z","type":"response_item","payload":{"type":"reasoning","summary":["Checking tool output"],"encrypted_content":"sealed"}}`,
-		`{"timestamp":"2026-05-18T10:00:03Z","type":"response_item","payload":{"type":"function_call","call_id":"call-1","name":"exec_command","arguments":{"cmd":"go test ./pkg/api"}}}`,
-		`{"timestamp":"2026-05-18T10:00:04Z","type":"response_item","payload":{"type":"function_call_output","call_id":"call-1","output":"ok","status":"completed"}}`,
-		`{"timestamp":"2026-05-18T10:00:05Z","type":"event_msg","payload":{"type":"agent_message","message":"The package tests passed."}}`,
-		`{"timestamp":"2026-05-18T10:00:06Z","type":"event_msg","payload":{"type":"task_started","message":"Applying follow-up patch"}}`,
-		`{"timestamp":"2026-05-18T10:00:07Z","type":"event_msg","payload":{"type":"agent_reasoning","text":"Need one more validation step."}}`,
-		`{"timestamp":"2026-05-18T10:00:08Z","type":"event_msg","payload":{"type":"new_future_event"}}`,
-		`{"timestamp":"2026-05-18T10:00:09Z","type":"unexpected_top_level"}`,
-		`{bad json`,
-	}, "\n")
-
-	parsed, err := parseCodexSessionDetails(strings.NewReader(session))
-	if err != nil {
-		t.Fatalf("parse codex session details: %v", err)
-	}
-
-	assertMixedCodexSessionSummary(t, parsed.Summary)
-	assertMixedCodexSessionTranscript(t, parsed)
-}
-
-func assertCodexSessionSummaryCoreCounts(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	if summary.LineCount != 10 || summary.EventCount != 9 || summary.MalformedLineCount != 1 || summary.UnknownEventCount != 2 || len(summary.Turns) != 2 || len(summary.FunctionCalls) != 2 {
-		t.Fatalf("summary = %#v, want parsed counts and two turns/calls", summary)
-	}
-}
-
-func assertCodexSessionSummaryFunctionCalls(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	firstCall := summary.FunctionCalls[0]
-	if firstCall.Order != 1 || stringValue(firstCall.Name) != "exec_command" || stringValue(firstCall.Arguments) != `{"cmd":"go test ./pkg/api"}` || stringValue(firstCall.Output) != "ok" || stringValue(firstCall.Status) != "completed" {
-		t.Fatalf("first function call = %#v, want completed exec_command call", firstCall)
-	}
-	secondCall := summary.FunctionCalls[1]
-	if secondCall.Order != 2 || stringValue(secondCall.Name) != "apply_patch" || stringValue(secondCall.Status) != "in_progress" || stringValue(secondCall.Output) != "" {
-		t.Fatalf("second function call = %#v, want in-progress custom tool call", secondCall)
-	}
-}
-
-func assertCodexSessionSummaryReasoning(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	if len(summary.Reasoning) != 1 || stringValue(summary.Reasoning[0].Summary) != `["checked input"]` || summary.Reasoning[0].Encrypted == nil || !*summary.Reasoning[0].Encrypted || stringValue(summary.Reasoning[0].EncryptedContent) != "sealed" {
-		t.Fatalf("reasoning = %#v, want summary, encrypted marker, and encrypted content", summary.Reasoning)
-	}
-}
-
-func assertParsedCodexSessionSummaryTranscript(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	if len(parsed.Transcript) != 4 {
-		t.Fatalf("transcript = %#v, want four ordered transcript entries", parsed.Transcript)
-	}
-	if parsed.Transcript[0].Type != factoryapi.Reasoning || stringValue(parsed.Transcript[0].SourceType) != "reasoning" || intValue(parsed.Transcript[0].LineNumber) != 2 {
-		t.Fatalf("first transcript entry = %#v, want reasoning line 2", parsed.Transcript[0])
-	}
-	if parsed.Transcript[1].Type != factoryapi.ToolCall || stringValue(parsed.Transcript[1].Name) != "exec_command" || stringValue(parsed.Transcript[1].Arguments) != `{"cmd":"go test ./pkg/api"}` {
-		t.Fatalf("second transcript entry = %#v, want exec_command tool call", parsed.Transcript[1])
-	}
-	if parsed.Transcript[2].Type != factoryapi.ToolOutput || stringValue(parsed.Transcript[2].Output) != "ok" || stringValue(parsed.Transcript[2].Status) != "completed" {
-		t.Fatalf("third transcript entry = %#v, want completed tool output", parsed.Transcript[2])
-	}
-	if parsed.Transcript[3].Type != factoryapi.ToolCall || stringValue(parsed.Transcript[3].Name) != "apply_patch" || stringValue(parsed.Transcript[3].Status) != "in_progress" {
-		t.Fatalf("fourth transcript entry = %#v, want in-progress apply_patch tool call", parsed.Transcript[3])
-	}
-	if parsed.Transcript[3].Order != 4 {
-		t.Fatalf("final transcript entry order = %d, want 4", parsed.Transcript[3].Order)
-	}
-}
-
-func assertCodexSessionSummaryTokenUsage(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	if summary.TokenUsage == nil || intValue(summary.TokenUsage.InputTokens) != 100 || intValue(summary.TokenUsage.CachedInputTokens) != 40 || intValue(summary.TokenUsage.OutputTokens) != 25 || intValue(summary.TokenUsage.ReasoningOutputTokens) != 5 || intValue(summary.TokenUsage.TotalTokens) != 130 {
-		t.Fatalf("token usage = %#v, want total consumed token fields", summary.TokenUsage)
-	}
-}
-
-func assertCodexSessionSummaryUnknowns(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	if len(summary.UnknownEvents) != 2 || summary.UnknownEvents[0].LineNumber != 8 || stringValue(summary.UnknownEvents[0].Type) != "event_msg" || stringValue(summary.UnknownEvents[0].PayloadType) != "new_future_event" || summary.UnknownEvents[1].LineNumber != 9 || stringValue(summary.UnknownEvents[1].Type) != "unexpected_top_level" {
-		t.Fatalf("unknown events = %#v, want compact line-level unknown records", summary.UnknownEvents)
-	}
-	if len(summary.ParseErrors) != 1 || summary.ParseErrors[0].LineNumber != 10 {
-		t.Fatalf("parse errors = %#v, want malformed line retained", summary.ParseErrors)
-	}
-}
-
-func assertMixedCodexSessionSummary(t *testing.T, summary factoryapi.ProviderSessionParseSummary) {
-	t.Helper()
-
-	if summary.LineCount != 11 || summary.EventCount != 10 || summary.MalformedLineCount != 1 || summary.UnknownEventCount != 2 {
-		t.Fatalf("summary = %#v, want mixed-session diagnostic counts", summary)
-	}
-	if len(summary.Turns) != 1 || summary.Turns[0].FunctionCallCount != 1 || summary.Turns[0].ReasoningCount != 2 {
-		t.Fatalf("turn summary = %#v, want one turn with function and reasoning counts", summary.Turns)
-	}
-	if len(summary.UnknownEvents) != 2 || summary.UnknownEvents[0].LineNumber != 9 || summary.UnknownEvents[1].LineNumber != 10 {
-		t.Fatalf("unknown events = %#v, want unknown event_msg and top-level event retained", summary.UnknownEvents)
-	}
-	if len(summary.ParseErrors) != 1 || summary.ParseErrors[0].LineNumber != 11 {
-		t.Fatalf("parse errors = %#v, want malformed line 11 retained", summary.ParseErrors)
-	}
-}
-
-func assertMixedCodexSessionTranscriptLength(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	if len(parsed.Transcript) != 7 {
-		t.Fatalf("transcript = %#v, want seven ordered transcript entries", parsed.Transcript)
-	}
-}
-
-func assertMixedCodexSessionTranscriptEntry(t *testing.T, parsed parsedCodexSessionDetails, index int, wantType factoryapi.ProviderSessionTranscriptEntryType, wantLine int, wantText string) {
-	t.Helper()
-
-	entry := parsed.Transcript[index]
-	if entry.Order != index+1 || entry.Type != wantType || intValue(entry.LineNumber) != wantLine || stringValue(entry.Text) != wantText {
-		t.Fatalf("transcript[%d] = %#v, want order=%d type=%q line=%d text=%q", index, entry, index+1, wantType, wantLine, wantText)
-	}
-	if intValue(entry.TurnIndex) != 1 {
-		t.Fatalf("transcript[%d] turn index = %#v, want 1", index, entry.TurnIndex)
-	}
-}
-
-func assertMixedCodexSessionTranscript(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	assertMixedCodexSessionTranscriptLength(t, parsed)
-	assertMixedCodexSessionTranscriptUserMessage(t, parsed)
-	assertMixedCodexSessionTranscriptReasoning(t, parsed)
-	assertMixedCodexSessionTranscriptToolEvents(t, parsed)
-	assertMixedCodexSessionTranscriptAssistantAndSystemEvents(t, parsed)
-}
-
-func assertMixedCodexSessionTranscriptUserMessage(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	assertMixedCodexSessionTranscriptEntry(t, parsed, 0, factoryapi.UserMessage, 2, "Inspect the failing run.")
-	if parsed.Transcript[0].SourceType == nil || *parsed.Transcript[0].SourceType != "message" {
-		t.Fatalf("first transcript source type = %#v, want message", parsed.Transcript[0].SourceType)
-	}
-}
-
-func assertMixedCodexSessionTranscriptReasoning(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	if parsed.Transcript[1].Order != 2 || parsed.Transcript[1].Type != factoryapi.Reasoning || intValue(parsed.Transcript[1].LineNumber) != 3 || stringValue(parsed.Transcript[1].Summary) != `["Checking tool output"]` || parsed.Transcript[1].Encrypted == nil || !*parsed.Transcript[1].Encrypted || stringValue(parsed.Transcript[1].EncryptedContent) != "sealed" {
-		t.Fatalf("transcript[1] = %#v, want encrypted reasoning summary and content on line 3", parsed.Transcript[1])
-	}
-
-	assertMixedCodexSessionTranscriptEntry(t, parsed, 6, factoryapi.Reasoning, 8, "Need one more validation step.")
-	if parsed.Transcript[6].SourceType == nil || *parsed.Transcript[6].SourceType != "agent_reasoning" {
-		t.Fatalf("final reasoning transcript source type = %#v, want agent_reasoning", parsed.Transcript[6].SourceType)
-	}
-}
-
-func assertMixedCodexSessionTranscriptToolEvents(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	if parsed.Transcript[2].Order != 3 || parsed.Transcript[2].Type != factoryapi.ToolCall || intValue(parsed.Transcript[2].LineNumber) != 4 || stringValue(parsed.Transcript[2].Name) != "exec_command" {
-		t.Fatalf("transcript[2] = %#v, want tool call on line 4", parsed.Transcript[2])
-	}
-	if parsed.Transcript[3].Order != 4 || parsed.Transcript[3].Type != factoryapi.ToolOutput || intValue(parsed.Transcript[3].LineNumber) != 5 || stringValue(parsed.Transcript[3].Output) != "ok" || stringValue(parsed.Transcript[3].Status) != "completed" {
-		t.Fatalf("transcript[3] = %#v, want tool output on line 5", parsed.Transcript[3])
-	}
-}
-
-func assertMixedCodexSessionTranscriptAssistantAndSystemEvents(t *testing.T, parsed parsedCodexSessionDetails) {
-	t.Helper()
-
-	assertMixedCodexSessionTranscriptEntry(t, parsed, 4, factoryapi.AssistantMessage, 6, "The package tests passed.")
-	if parsed.Transcript[4].SourceType == nil || *parsed.Transcript[4].SourceType != "agent_message" {
-		t.Fatalf("assistant transcript source type = %#v, want agent_message", parsed.Transcript[4].SourceType)
-	}
-
-	assertMixedCodexSessionTranscriptEntry(t, parsed, 5, factoryapi.SystemEvent, 7, "Applying follow-up patch")
-	if parsed.Transcript[5].SourceType == nil || *parsed.Transcript[5].SourceType != "task_started" {
-		t.Fatalf("system-event transcript source type = %#v, want task_started", parsed.Transcript[5].SourceType)
-	}
-}
-
-func TestParseCodexSessionSummary_AcceptsLargeJSONLRecords(t *testing.T) {
-	session := strings.Join([]string{
-		`{"timestamp":"2026-05-18T10:00:00Z","type":"turn_context"}`,
-		`{"timestamp":"2026-05-18T10:00:01Z","type":"response_item","payload":{"type":"reasoning","content":"` + strings.Repeat("x", 128*1024) + `"}}`,
-	}, "\n")
-
-	summary, err := parseCodexSessionSummary(strings.NewReader(session))
-	if err != nil {
-		t.Fatalf("parse codex session summary: %v", err)
-	}
-	if summary.LineCount != 2 || summary.EventCount != 2 || len(summary.Reasoning) != 1 {
-		t.Fatalf("summary = %#v, want large response item parsed successfully", summary)
-	}
-}
-
-func TestGetProviderSessionDetails_NotFoundIsDistinguishable(t *testing.T) {
-	srv := newTestServerWithCodexRoot(t.TempDir())
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=codex&kind=session_id&id=missing-session", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusNotFound, "NOT_FOUND", "provider session not found")
-}
-
-func TestGetProviderSessionDetails_CursorNotFoundIsDistinguishable(t *testing.T) {
-	srv := newTestServerWithCursorRoot(t.TempDir())
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=cursor&kind=session_id&id=missing-session", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusNotFound, "NOT_FOUND", "provider session not found")
-}
-
-func TestGetProviderSessionDetails_LegacyAgentCursorNotFoundIsDistinguishable(t *testing.T) {
-	srv := newTestServerWithCursorRoot(t.TempDir())
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=agent&kind=session_id&id=missing-session", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusNotFound, "NOT_FOUND", "provider session not found")
-}
-
-func TestGetProviderSessionDetails_CursorNotFoundLogsDiagnostic(t *testing.T) {
-	root := t.TempDir()
-	core, logs := observer.New(zap.InfoLevel)
-	srv := NewServerWithOptions(&testutil.MockFactory{
-		Marking: &petri.MarkingSnapshot{
-			Tokens: make(map[string]*interfaces.Token),
-		},
-	}, 8080, zap.New(core), ServerOptions{CursorSessionsRoot: root})
-
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=cursor&kind=session_id&id=missing-session", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-
-	assertJSONError(t, rec, http.StatusNotFound, "NOT_FOUND", "provider session not found")
-
-	entries := logs.FilterMessage("cursor provider session lookup not found").AllUntimed()
-	if len(entries) != 1 {
-		t.Fatalf("cursor not-found diagnostic count = %d, want 1", len(entries))
-	}
-	fields := entries[0].ContextMap()
-	if fields["provider"] != "cursor" {
-		t.Fatalf("provider field = %#v, want cursor", fields["provider"])
-	}
-	if fields["lookup_kind"] != "session_id" {
-		t.Fatalf("lookup_kind field = %#v, want session_id", fields["lookup_kind"])
-	}
-	if fields["requested_id"] != "missing-session" {
-		t.Fatalf("requested_id field = %#v, want missing-session", fields["requested_id"])
-	}
-	if fields["searched_root"] != root {
-		t.Fatalf("searched_root field = %#v, want %q", fields["searched_root"], root)
-	}
-	if fields["root_configured"] != true {
-		t.Fatalf("root_configured field = %#v, want true", fields["root_configured"])
-	}
-}
-
-func TestGetProviderSessionDetails_IgnoresUnsupportedRolloutFileNames(t *testing.T) {
-	root := t.TempDir()
-	writeNamedProviderSessionFixture(t, root, "rollout-backup-sess_123.jsonl", `{"type":"session_meta","id":"sess_123"}`)
-	writeNamedProviderSessionFixture(t, root, "rollout-2026-05-20T17-35-24-backup-sess_123.jsonl", `{"type":"session_meta","id":"sess_123"}`)
-
-	srv := newTestServerWithCodexRoot(root)
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=codex&kind=session_id&id=sess_123", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusNotFound, "NOT_FOUND", "provider session not found")
-}
-
-func TestGetProviderSessionDetails_RejectsPathLikeAndMalformedIdentifiers(t *testing.T) {
-	for _, target := range []string{
-		"/provider-sessions/detail?provider=codex&kind=session_id&id=../secret",
-		"/provider-sessions/detail?provider=codex&kind=session_id&id=/tmp/rollout-session.jsonl",
-		"/provider-sessions/detail?provider=codex&kind=session_id&id=session.with.dot",
-	} {
-		t.Run(target, func(t *testing.T) {
-			srv := newTestServerWithCodexRoot(t.TempDir())
-			req := httptest.NewRequest("GET", target, nil)
-			rec := httptest.NewRecorder()
-			srv.Handler().ServeHTTP(rec, req)
-			assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "provider session must be a codex session_id identifier without path separators")
-		})
-	}
-}
-
-func TestGetProviderSessionDetails_CursorRejectsPathLikeAndMalformedIdentifiers(t *testing.T) {
-	for _, target := range []string{
-		"/provider-sessions/detail?provider=cursor&kind=session_id&id=../secret",
-		"/provider-sessions/detail?provider=cursor&kind=session_id&id=/tmp/store.db",
-		"/provider-sessions/detail?provider=cursor&kind=session_id&id=session.with.dot",
-	} {
-		t.Run(target, func(t *testing.T) {
-			srv := newTestServerWithCursorRoot(t.TempDir())
-			req := httptest.NewRequest("GET", target, nil)
-			rec := httptest.NewRecorder()
-			srv.Handler().ServeHTTP(rec, req)
-			assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "provider session must be a cursor session_id identifier without path separators")
-		})
-	}
-}
-
-func TestGetProviderSessionDetails_LegacyAgentCursorRejectsPathLikeAndMalformedIdentifiers(t *testing.T) {
-	for _, target := range []string{
-		"/provider-sessions/detail?provider=agent&kind=session_id&id=../secret",
-		"/provider-sessions/detail?provider=agent&kind=session_id&id=/tmp/store.db",
-		"/provider-sessions/detail?provider=agent&kind=session_id&id=session.with.dot",
-	} {
-		t.Run(target, func(t *testing.T) {
-			srv := newTestServerWithCursorRoot(t.TempDir())
-			req := httptest.NewRequest("GET", target, nil)
-			rec := httptest.NewRecorder()
-			srv.Handler().ServeHTTP(rec, req)
-			assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "provider session must be a cursor session_id identifier without path separators")
-		})
-	}
-}
-
-func TestGetProviderSessionDetails_RejectsUnsupportedProviderOrKindByContract(t *testing.T) {
-	for _, target := range []string{
-		"/provider-sessions/detail?provider=openai&kind=session_id&id=sess-123",
-		"/provider-sessions/detail?provider=codex&kind=path&id=sess-123",
-		"/provider-sessions/detail?provider=cursor&kind=path&id=sess-123",
-	} {
-		t.Run(target, func(t *testing.T) {
-			srv := newTestServerWithCodexRoot(t.TempDir())
-			req := httptest.NewRequest("GET", target, nil)
-			rec := httptest.NewRecorder()
-			srv.Handler().ServeHTTP(rec, req)
-			assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "invalid request parameter")
-		})
-	}
-}
-
-func TestGetProviderSessionDetails_RejectsSessionSymlinkOutsideConfiguredRoot(t *testing.T) {
-	root := t.TempDir()
-	outside := t.TempDir()
-	outsideSessionPath := filepath.Join(outside, "rollout-sess-outside.jsonl")
-	if err := os.WriteFile(outsideSessionPath, []byte(`{"type":"session_meta"}`), 0o600); err != nil {
-		t.Fatalf("write outside session fixture: %v", err)
-	}
-	sessionDir := filepath.Join(root, "2026", "05", "18")
-	if err := os.MkdirAll(sessionDir, 0o755); err != nil {
-		t.Fatalf("create session dir: %v", err)
-	}
-	if err := os.Symlink(outsideSessionPath, filepath.Join(sessionDir, "rollout-sess-outside.jsonl")); err != nil {
-		t.Fatalf("create provider session symlink: %v", err)
-	}
-
-	srv := newTestServerWithCodexRoot(root)
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=codex&kind=session_id&id=sess-outside", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "provider session must be a codex session_id identifier without path separators")
-}
-
-func TestGetProviderSessionDetails_RejectsSessionSymlinkOutsideConfiguredRootEvenWhenValidMatchExists(t *testing.T) {
-	root := t.TempDir()
-	writeProviderSessionFixture(t, root, "sess-shared", `{"type":"session_meta","id":"sess-shared"}`)
-	outside := t.TempDir()
-	outsideSessionPath := filepath.Join(outside, "rollout-2026-05-20T17-35-24-sess-shared.jsonl")
-	if err := os.WriteFile(outsideSessionPath, []byte(`{"type":"session_meta"}`), 0o600); err != nil {
-		t.Fatalf("write outside session fixture: %v", err)
-	}
-	sessionDir := filepath.Join(root, "2026", "05", "18")
-	if err := os.Symlink(outsideSessionPath, filepath.Join(sessionDir, "rollout-2026-05-20T17-35-24-sess-shared.jsonl")); err != nil {
-		t.Fatalf("create provider session symlink: %v", err)
-	}
-
-	srv := newTestServerWithCodexRoot(root)
-	req := httptest.NewRequest("GET", "/provider-sessions/detail?provider=codex&kind=session_id&id=sess-shared", nil)
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-	assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", "provider session must be a codex session_id identifier without path separators")
-}
